% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_documentation.R
\docType{data}
\name{R83P1}
\alias{R83P1}
\title{Table R83P1 \cr
Member characteristics of rural household in 1383 \cr
خصوصيات اجتماعي اعضاي خانوار روستایی سال 1383}
\format{
A data frame with 10 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{آدرس}
\item Description: Household address
}
\item \code{COL01}:
\itemize{
\item Farsi name: \emph{شماره ردیف اعضای خانوار}
\item Description: Row number of family members
}
\item \code{COL03}:
\itemize{
\item Farsi name: \emph{بستگی با سرپرست}
\item Description: Relationship with the guardian
}
\item \code{COL04}:
\itemize{
\item Farsi name: \emph{جنس}
\item Description: Gender
}
\item \code{COL05}:
\itemize{
\item Farsi name: \emph{سن}
\item Description: Age
}
\item \code{COL06}:
\itemize{
\item Farsi name: \emph{وضع سواد}
\item Description: Literacy status
}
\item \code{COL07}:
\itemize{
\item Farsi name: \emph{آیا در حال حاضر تحصیل می کند؟}
\item Description: Is she/he currently studying?
}
\item \code{COL08}:
\itemize{
\item Farsi name: \emph{پایه یا مدرک}
\item Description: Education degree
}
\item \code{COL09}:
\itemize{
\item Farsi name: \emph{وضع فعالیت}
\item Description: Employment status
}
\item \code{COL10}:
\itemize{
\item Farsi name: \emph{وضع زناشویی}
\item Description: Marital status
}
}
}
\usage{
R83P1
}
\description{
{R for Rural} \cr
{83 for 1383} \cr
{P for Part}
}
\keyword{datasets}
