% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_documentation.R
\docType{data}
\name{R83P2}
\alias{R83P2}
\title{Table R83P2 \cr
Rural household Facilities in 1383 \cr
مشخصات محل سكونت زندگي روستایی سال 1383}
\format{
A data frame with 37 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{آدرس}
\item Description: Household address
}
\item \code{DYCOL01}:
\itemize{
\item Farsi name: \emph{نحوه ی تصرف محل سکونت}
\item Description: tenure
}
\item \code{DYCOL02}:
\itemize{
\item Farsi name: \emph{تعداد اتاق در اختيار خانوار}
\item Description: Room
}
\item \code{DYCOL03}:
\itemize{
\item Farsi name: \emph{سطح زيربنا}
\item Description: Space
}
\item \code{DYCOL04}:
\itemize{
\item Farsi name: \emph{نوع اسكلت بناي محل سكونت}
\item Description: Construction
}
\item \code{DYCOL05}:
\itemize{
\item Farsi name: \emph{مصالح عمده}
\item Description: Material
}
\item \code{DYCOL06}:
\itemize{
\item Farsi name: \emph{اتومبيل سواري شخصي}
\item Description: Vehicle
}
\item \code{DYCOL07}:
\itemize{
\item Farsi name: \emph{موتورسيكلت}
\item Description: Motorcycle
}
\item \code{DYCOL08}:
\itemize{
\item Farsi name: \emph{دوچرخه}
\item Description: Bicycle
}
\item \code{DYCOL09}:
\itemize{
\item Farsi name: \emph{راديو}
\item Description: Radio
}
\item \code{DYCOL10}:
\itemize{
\item Farsi name: \emph{راديو ضبط، ضبط و پخش صوت}
\item Description: Radio tape
}
\item \code{DYCOL11}:
\itemize{
\item Farsi name: \emph{تلوزيون (سياه وسفيد)}
\item Description: TV (black and white)
}
\item \code{DYCOL12}:
\itemize{
\item Farsi name: \emph{تلويزيون رنگي}
\item Description: TV
}
\item \code{DYCOL13}:
\itemize{
\item Farsi name: \emph{VCD و DVD انواع ویذیو و}
\item Description: VHS_VCD_DVD
}
\item \code{DYCOL14}:
\itemize{
\item Farsi name: \emph{رایانه}
\item Description: Computer
}
\item \code{DYCOL15}:
\itemize{
\item Farsi name: \emph{تلفن همراه (غير شغلي)}
\item Description: Cellphone
}
\item \code{DYCOL16}:
\itemize{
\item Farsi name: \emph{فريزر}
\item Description: Freezer
}
\item \code{DYCOL17}:
\itemize{
\item Farsi name: \emph{يخچال}
\item Description: Refrigerator
}
\item \code{DYCOL18}:
\itemize{
\item Farsi name: \emph{يخچال فريزر}
\item Description: Fridge
}
\item \code{DYCOL19}:
\itemize{
\item Farsi name: \emph{اجاق گاز}
\item Description: Stove
}
\item \code{DYCOL20}:
\itemize{
\item Farsi name: \emph{جاروبرقي}
\item Description: Vacuum
}
\item \code{DYCOL21}:
\itemize{
\item Farsi name: \emph{ماشین لباسشویی}
\item Description: Washing machine
}
\item \code{DYCOL22}:
\itemize{
\item Farsi name: \emph{چرخ خياطي}
\item Description: Sewingmachine
}
\item \code{DYCOL23}:
\itemize{
\item Farsi name: \emph{پنکه}
\item Description: Fan
}
\item \code{DYCOL24}:
\itemize{
\item Farsi name: \emph{کولر آبي متحرک}
\item Description: Evap cooling portable
}
\item \code{DYCOL25}:
\itemize{
\item Farsi name: \emph{كولر گازي متحرك}
\item Description: Split portable
}
\item \code{DYCOL26}:
\itemize{
\item Farsi name: \emph{ماشين ظرفشويي}
\item Description: Dishwasher
}
\item \code{DYCOL27}:
\itemize{
\item Farsi name: \emph{هيچكدام}
\item Description: None
}
\item \code{DYCOL28}:
\itemize{
\item Farsi name: \emph{آب لوله كشي}
\item Description: Pipewater
}
\item \code{DYCOL29}:
\itemize{
\item Farsi name: \emph{برق}
\item Description: Electricity
}
\item \code{DYCOL30}:
\itemize{
\item Farsi name: \emph{گازلوله كشي}
\item Description: Pipegas
}
\item \code{DYCOL31}:
\itemize{
\item Farsi name: \emph{تلفن}
\item Description: Telephone
}
\item \code{DYCOL32}:
\itemize{
\item Farsi name: \emph{اينترنت}
\item Description: Internet
}
\item \code{DYCOL33}:
\itemize{
\item Farsi name: \emph{حمام}
\item Description: Bathroom
}
\item \code{DYCOL34}:
\itemize{
\item Farsi name: \emph{آشپزخانه}
\item Description: Kitchen
}
\item \code{DYCOL35}:
\itemize{
\item Farsi name: \emph{كولرآبي ثابت}
\item Description: Evap cooling
}
\item \code{DYCOL36}:
\itemize{
\item Farsi name: \emph{برودت مرکزی}
\item Description: Central cooling
}
}
}
\usage{
R83P2
}
\description{
{R for Rural} \cr
{83 for 1383} \cr
{P for Part}
}
\keyword{datasets}
