% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documentation.R
\docType{data}
\name{R83P3S01}
\alias{R83P3S01}
\title{Table R83P3S01 \cr
Household food and tobacco expenditure of rural household in 1383 \cr
هزينه هاي خوراكي و دخانی خانوار روستایی سال 1383}
\format{
A data frame with 7 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{آدرس}
\item Description: Household address
}
\item \code{COL1}:
\itemize{
\item Farsi name: \emph{کد کالا}
\item Description: Bar code
}
\item \code{COL3}:
\itemize{
\item Farsi name: \emph {اطلاعاتی ثبت نشده است}
\item Description: No description
}
\item \code{COL4}:
\itemize{
\item Farsi name: \emph {اطلاعاتی ثبت نشده است}
\item Description: No description
}
\item \code{COL5}:
\itemize{
\item Farsi name: \emph {اطلاعاتی ثبت نشده است}
\item Description: No description
}
\item \code{COL6}:
\itemize{
\item Farsi name: \emph {اطلاعاتی ثبت نشده است}
\item Description: No description
}
\item \code{COL7}:
\itemize{
\item Farsi name: \emph {اطلاعاتی ثبت نشده است}
\item Description: No description
}
}
}
\usage{
R83P3S01
}
\description{
{R for Rural} \cr
{83 for 1383} \cr
{P for Part} \cr
{S for Section}
}
\keyword{datasets}
