% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documentation.R
\docType{data}
\name{R83P3S10}
\alias{R83P3S10}
\title{Table R83P3S10 \cr
Household recreation & leisure expenditure of rural household in 1383 \cr
هزينه هاي  مربوط به خدمات فرهنگي و تفريحات خانوار روستایی سال 1383}
\format{
A data frame with 4 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{آدرس خانوار}
\item Description: Household address
}
\item \code{COL1}:
\itemize{
\item Farsi name: \emph{کد کالا}
\item Description: Bar code
}
\item \code{COL3}:
\itemize{
\item Farsi name: \emph{طریق تهیه}
\item Description: Preparation method
}
\item \code{COL4}:
\itemize{
\item Farsi name: \emph{ارزش (ریال)}
\item Description: Value (Rials)
}
}
}
\usage{
R83P3S10
}
\description{
{R for Rural} \cr
{83 for 1383} \cr
{P for Part} \cr
{S for Section}
}
\keyword{datasets}
