% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documentation.R
\docType{data}
\name{R83P3S13}
\alias{R83P3S13}
\title{Table R83P3S13 \cr
Supply and sale of durable goods & other yearly expenditure of rural household in 1383 \cr
تهيه و فروش كالاهای بادوام منزل روستایی سال 1383}
\format{
A data frame with 5 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{آدرس خانوار}
\item Description: Household address
}
\item \code{COL1}:
\itemize{
\item Farsi name: \emph{کد کالا}
\item Description: Bar code
}
\item \code{COL3}:
\itemize{
\item Farsi name: \emph{طریق تهیه}
\item Description: Preparation method
}
\item \code{COL4}:
\itemize{
\item Farsi name: \emph{خرید یا هزینه}
\item Description: Purchase or expenditure
}
\item \code{COL5}:
\itemize{
\item Farsi name: \emph{فروش دست دوم}
\item Description: Second hand sale
}
}
}
\usage{
R83P3S13
}
\description{
{R for Rural} \cr
{83 for 1383} \cr
{P for Part} \cr
{S for Section}
}
\keyword{datasets}
