% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documentation.R
\docType{data}
\name{R83P4S2}
\alias{R83P4S2}
\title{Table R83P4S2 \cr
Self-employment income of rural household in 1383 \cr
درآمد پولي اعضای شاغل خانوار روستایی از مشاغل آزاد سال 1383}
\format{
A data frame with 14 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{آدرس}
\item Description: Household address
}
\item \code{COL01}:
\itemize{
\item Farsi name: \emph{شماره ردیف عضو شاغل}
\item Description: Row number of the working member
}
\item \code{COL02}:
\itemize{
\item Farsi name: \emph{آيا در حال حاضر شاغل است؟}
\item Description: Is she/he currently employed?
}
\item \code{COL03}:
\itemize{
\item Farsi name: \emph{كدشغل}
\item Description: Job code
}
\item \code{COL04}:
\itemize{
\item Farsi name: \emph{فعاليت اصلي محل كار}
\item Description: Main activity of the workplace
}
\item \code{COL05}:
\itemize{
\item Farsi name: \emph{کارفرما 1 کارکن مستقل 2 کارکن فامیلی 3}
\item Description: Employer 1 independent employee 2 family employee 3
}
\item \code{COL06}:
\itemize{
\item Farsi name: \emph{كد شغل کشاورزی 1 غیرکشاورزی 2}
\item Description: Agriculture 1 non-agriculture 2
}
\item \code{COL07}:
\itemize{
\item Farsi name: \emph{ميزان ساعت هاي كار در روز}
\item Description: The amount of working hours per day
}
\item \code{COL08}:
\itemize{
\item Farsi name: \emph{تعداد روزهاي كار در هفته}
\item Description: Number of working days per week
}
\item \code{COL09}:
\itemize{
\item Farsi name: \emph{مزد و حقوق و مزایا}
\item Description: Wages and benefits
}
\item \code{COL10}:
\itemize{
\item Farsi name: \emph{بذر، آب و کود}
\item Description: Seeds, water and fertilizer
}
\item \code{COL11}:
\itemize{
\item Farsi name: \emph{تهیه ابزار کار بی دوام}
\item Description: Preparation of unsustainable working tools
}
\item \code{COL12}:
\itemize{
\item Farsi name: \emph{کارمزد شغلی}
\item Description: Job wage
}
\item \code{COL13}:
\itemize{
\item Farsi name: \emph{مالیات شغلی}
\item Description: Job tax
}
}
}
\usage{
R83P4S2
}
\description{
{R for Rural} \cr
{83 for 1383} \cr
{P for Part} \cr
{S for Section}
}
\keyword{datasets}
