% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documentation.R
\docType{data}
\name{R83P4S3}
\alias{R83P4S3}
\title{Table R83P4S3 \cr
Miscellaneous income of rural household in 1383 \cr
درآمد هاي متفرقه خانوار روستایی سال 1383}
\format{
A data frame with 8 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{آدرس}
\item Description: Household address
}
\item \code{COL01}:
\itemize{
\item Farsi name: \emph{شماره ردیف اعضای خانوار}
\item Description: Row number of family members
}
\item \code{COL03}:
\itemize{
\item Farsi name: \emph{حقوق بازنشستگی}
\item Description: Retirement wage
}
\item \code{COL04}:
\itemize{
\item Farsi name: \emph{درآمد حاصل از اجاره محل کسب}
\item Description: Revenue from renting the business place
}
\item \code{COL05}:
\itemize{
\item Farsi name: \emph{درآمد حاصل از حساب پس انداز}
\item Description: Income from the savings account
}
\item \code{COL06}:
\itemize{
\item Farsi name: \emph{کمک هزینه تحصیلی}
\item Description: Scholarships
}
\item \code{COL07}:
\itemize{
\item Farsi name: \emph{درآمد حاصل از محل فروش مصنوعات}
\item Description: Income from the sale of artifacts
}
\item \code{COL08}:
\itemize{
\item Farsi name: \emph{دریافتی های انتقالی از خانوارهای دیگر}
\item Description: Transfer receivables from other families
}
}
}
\usage{
R83P4S3
}
\description{
{R for Rural} \cr
{83 for 1383} \cr
{P for Part} \cr
{S for Section}
}
\keyword{datasets}
