% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documentation.R
\docType{data}
\name{U83P3S02}
\alias{U83P3S02}
\title{Table U83P3S02 \cr
Expenditure of unclassified beverages of urban household in 1383 \cr
هزينه هاي نوشيدني های طبقه بندی نشده خانوار شهری سال 1383}
\format{
A data frame with 7 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{آدرس}
\item Description: Household address
}
\item \code{DYCOL01}:
\itemize{
\item Farsi name: \emph{کد کالا}
\item Description: Bar code
}
\item \code{COL3}:
\itemize{
\item Farsi name: \emph
\item Description: No description
}
\item \code{COL4}:
\itemize{
\item Farsi name: \emph
\item Description: No description
}
\item \code{COL5}:
\itemize{
\item Farsi name: \emph
\item Description: No description
}
\item \code{COL6}:
\itemize{
\item Farsi name: \emph
\item Description: No description
}
\item \code{COL7}:
\itemize{
\item Farsi name: \emph
\item Description: No description
}
}
}
\usage{
U83P3S02
}
\description{
{U for Urban} \cr
{83 for 1383} \cr
{P for Part} \cr
{S for Section}
}
\keyword{datasets}
