% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documentation.R
\docType{data}
\name{U83P3S14}
\alias{U83P3S14}
\title{Table U83P3S14 \cr
Household investment of urban household in the last 12 months in 1383 \cr
سرمايه گذاري خانوار شهری در 12 ماه گذشته سال 1383}
\format{
A data frame with 5 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{آدرس خانوار}
\item Description: Household address
}
\item \code{COL1}:
\itemize{
\item Farsi name: \emph{کد کالا}
\item Description: Bar code
}
\item \code{COL3}:
\itemize{
\item Farsi name: \emph{طریق تهیه}
\item Description: Preparation method
}
\item \code{COL4}:
\itemize{
\item Farsi name: \emph{خرید یا هزینه}
\item Description: Purchase or expenditure
}
\item \code{COL5}:
\itemize{
\item Farsi name: \emph{فروش دست دوم}
\item Description: Second hand sale
}
}
}
\usage{
U83P3S14
}
\description{
{U for Urban} \cr
{83 for 1383} \cr
{P for Part} \cr
{S for Section}
}
\keyword{datasets}
